refactor subscription code
