+++ src/plugins/subscriptions/helpers.js
@@ -1,0 +1,7 @@
+export const omit = (prop: string, obj = {}) =>
+  Object.keys(obj).reduce((next, key: string) => {
+    if (key !== prop) {
+      next[key] = obj[key]
+    }
+    return next
+  }, {})
